+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Version 1.x.y - YYYY-MM-DD
 
 - Implement ``do_ping`` for GSheets dialect
 - Create a ``cookiecutter`` template for new adapters
+- Add a ``StringDuration`` field
 
 Version 1.0.4 - 2021-08-30
 ==========================